+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.2 - November 2024
 
+- incompatible change: `term:print` was renamed to `term:print-styled`
 - added vspace, banner
 - added vtable
 - table can now print lists of plists, lists of hash-tables, single elements.